test: add tests for popout button
